+++ pallets/governance/src/benchmarking.rs
@@ -129,7 +129,8 @@ benchmarks! {
             params.min_validator_stake,
             params.max_allowed_validators,
             params.use_weights_encryption,
-            params.copier_margin
+            params.copier_margin,
+            params.max_encryption_period
         )?;
 
         // add balance to submit the proposal
@@ -161,7 +162,8 @@ benchmarks! {
         params.min_validator_stake,
         params.max_allowed_validators,
         params.use_weights_encryption,
-        params.copier_margin
+        params.copier_margin,
+        params.max_encryption_period
     )
 
     // 2