+++ pallets/subspace/src/selections/dispatches.rs
@@ -131,6 +131,7 @@ pub mod dispatches {
             max_allowed_validators: Option<u16>,
             use_weights_encryption: bool,
             copier_margin: I64F64,
+            max_encryption_period: Option<u64>,
         ) -> DispatchResult {
             let params = SubnetParams {
                 founder,
@@ -155,6 +156,7 @@ pub mod dispatches {
                 metadata,
                 use_weights_encryption,
                 copier_margin,
+                max_encryption_period,
             };
 
             let changeset = SubnetChangeset::update(netuid, params)?;